+++ tools/ci/build-ubuntu-with-latest-clang-gcc.sh
@@ -31,16 +31,6 @@ cd "${WORKSPACE}"
 msg "installing build dependencies"
 sudo apt-get update && sudo apt-get install -y libacl1-dev libncurses5-dev
 
-if [ "$COMPILER" == "clang" ]; then
-    msg "installing latest stable clang"
-    wget https://apt.llvm.org/llvm.sh -O /tmp/llvm.sh
-    chmod +x /tmp/llvm.sh
-    # set CURRENT_LLVM_STABLE
-    # shellcheck disable=SC2002
-    eval "$(cat /tmp/llvm.sh | grep CURRENT_LLVM_STABLE= -m 1)"
-    sudo /tmp/llvm.sh "${CURRENT_LLVM_STABLE}"
-fi
-
 msg "creating local test users and groups for testing access control"
 sudo ./tools/scripts/add_test_users.sh
 
@@ -52,7 +42,5 @@ if [ "$COMPILER" == "gcc" ]; then
 fi
 
 if [ "$COMPILER" == "clang" ]; then
-    export CC=clang-${CURRENT_LLVM_STABLE}
-    export CXX=clang++-${CURRENT_LLVM_STABLE}
     ./tools/iceoryx_build_test.sh clean clang release build-all out-of-tree build-shared test-add-user
 fi